Fix timezoning and overlapping time check
