+++ appsettings.json
@@ -9,7 +9,7 @@
   "GoogleApi":{
     "AuthBaseUrl":"https://accounts.google.com/o/oauth2/v2/auth",
     "RequestTokenUrl":"https://oauth2.googleapis.com/token",
-    "ClientID": "136668314012-d45m2upreh8jeroe2vn84kjm36ov77aa.apps.googleusercontent.com",
+    "ClientID": "",
     "RedirectUrl":"http://localhost:8080",
     "Scope":"https://www.googleapis.com/auth/calendar email openid", 
     "CalendarUrl": "https://www.googleapis.com/calendar/v3/calendars/primary/events"
